split test_app.py in half
